Set up signal handling to handle SIGKILL
Turns out that Plex kills the transcoder process with SIGKILL.
Since SIGKILL cannot be caught, we need to set up a wrapper that
launches the transcoder in a new process.
The downside is that we need to monitor if the parent process still
exists.

Also switches the old signal hangling to use context based signal
handling.