Add support for docker-pullable:// images
Docker has a special URI notation for image IDs, handle those image
types by removing the URI prefix.

This is fragile, but aside from talking to the node itself to find the
correct image ID, this seems to be the simplest solution. Details of the
issue are in kubernetes/kubernetes#95968

Fixes #11